Implemented Redux store for central state management and divided playground into seperate components
