+++ frontend/src/App.tsx
@@ -4,7 +4,6 @@ import Home from '@/pages/Home';
 import Login from './pages/Login';
 import SignUp from './pages/SignUp';
 import Dashboard from './pages/Dashboard';
-import ChatPage from './pages/ChatPage';
 import Playground from './pages/Playground';
 
 
@@ -17,7 +16,6 @@ function App() {
         <Route path="/login" element={<Login />} />
         <Route path="/signup" element={<SignUp />} />
         <Route path="/dashboard" element={<Dashboard />} />
-        <Route path="/dashboard/chat" element={<ChatPage />} />
         <Route path="/dashboard/playground" element={<Playground />} />
         {/* other routes */}
       </Routes>
