+++ frontend/src/main.tsx
@@ -1,10 +1,14 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
+import {Provider} from 'react-redux'
 import './index.css';
 import App from './App.tsx';
+import store from './store/index.ts';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+    <Provider store={store}>
+      <App />
+    </Provider>
   </StrictMode>
 );
